dbContext, seed the data, model fix

+++ WorkshopPro/Models/Employee.cs
@@ -6,7 +6,7 @@
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string Position { get; set; }
-        public decimal Salary { get; set; }
+        public int Salary { get; set; }
         public Workshop Workshop { get; set; }
         //public int? ProjectId { get; set; }
         public Project Project { get; set; }
